+++ HISTORY.md
@@ -2,6 +2,11 @@
 
 ## 0.39.0
 
+## 0.38.1
+
+Added `from_linked_vec_transform` and `from_vec_transform` methods for `ProductNamedTupleDistribution`.
+This patch allows sampling from `ProductNamedTupleDistribution` in DynamicPPL models.
+
 ## 0.38.0
 
 ### Breaking changes